refactor(frontend): replace base64 image handling with Cloudinary URLs
