+++ vietsneaker-ui/src/app/(page)/(main)/checkout/confirm/page.tsx
@@ -5,7 +5,7 @@ import React, { MouseEvent } from "react";
 import { formatVND } from "@/util/currency";
 import { OrderControllerService } from "@/gen";
 import { logger } from "@/util/logger";
-import { ProductImage } from "@/app/components/common/base-64-image";
+import { ProductImage } from "@/app/components/common/ProductImage";
 import Thumbnail from "@/app/components/common/thumbnail";
 import { useRouter } from "next/navigation";
 import { toast } from "react-toastify";
