+++ vietsneaker-ui/src/app/(page)/admin/(inventory)/purchases/components/purchase-details-table-row.tsx
@@ -6,7 +6,7 @@
 "use client";
 import React, { useState } from "react";
 import Thumbnail from "@/app/components/common/thumbnail";
-import { ProductImage } from "@/app/components/common/base-64-image";
+import { ProductImage } from "@/app/components/common/ProductImage";
 import { formatVND } from "@/util/currency";
 import { PurchaseItemDetails } from "@/gen";
