+++ vietsneaker-ui/src/app/(page)/admin/(inventory)/purchases/new/components/edit-purchase-item-modal.tsx
@@ -1,4 +1,4 @@
-import { ProductImage } from "@/app/components/common/base-64-image";
+import { ProductImage } from "@/app/components/common/ProductImage";
 import Thumbnail from "@/app/components/common/thumbnail";
 import ErrorText from "@/app/components/typography/error-text";
 import {
